Complete version of problem 4013
 Be aware of the index calculations. To ease debugging, it is
recommended to set variables separately for indexes that are calculated.
ex)curRightIndex, curLeftIndex used in this solution.

 Be aware of the input style. The one it took a lot of time for this
problem was that input was given as 1 ~ 4. Since I thought it is from 0
~ 3, rotation operation was done for excess memory too. This caused
fatal error for the magnets have been changed into some values that
shouldn't be of.
ex) In this case, the magnets got random values like 7, 0, -1, -2  etc.

 Also this means you would benefit from using inequality instead of
using equality.(line 93, 98)

 Think of the 순서, when designing the code. In this case, shold I
rotate the right one before I go into rotation calculation or not?

 c++ supports negative value initialization, that is, you can assign a
value in this format
 >> nextDirection = -curDirection;