[REVIEW] How to use binary search
 The toughest condition is when given 98% exactly. Even then, winning
exactly the same amount of total games played can suitably increment the
winrate by one. Therefore, the upperbound is not 1e10 but game_total
from the input.
[line 16]

 You can't increment the winrate if the original win_rate was gte 99.
print -1 in this case.[line 10]

 You need to find when the rightmost equal winrate ends. That's why you
increment low by one[line 24]. When the winrate_new is gt
winrate_original, the amount should get reduced and we need high = mid.
High must not surpass the mid value or it can lead to catastrophic
problem.[line 27]