[REVIEW] Weekly Contest 161
1. How to make 2 different strings of xy into the same by swapping.

2. Count number of nice subarrays. You save the indexes odd integers
inside a list. If given k, you should get subarray that has k odd
integers. To implement this, you get k + 1 inputs inside the list, get
the possibles.
ex) if k is 2 and you have
0 0 1 0 0 1 0 0 0 1
Then you have list [-1, 2, 5, 9, 10]. -1, 10 is automatically created to
imply the start and end. Then, you can get
3 * 4 + 3 * 1 as an aswer.

3. Minimum Remove to Make Valid Parenthesis.
 They say it's a typical problem. You must store the 'index' of the left
parenthesis of the string into the stack.
 If right parenthesis is given, check if any matching left parentheses
exists. If it does, you need to do nothing. If there isn't, you must
remove this right parenthesis.
 In solution, it makes the character into '*' which is some character
that means to remove. Then read the string again and reject when * is
read.
 After all string is read, check if there're any left parentheses left
inside the stack. If there are, you must remove them too.
 I implemented this with using priority queue, but maybe using vector
 and sorting it show better performance.

4. Check if it is good array.

 Make gcd of the first 2 values everytime you meet the new array. If you
get 1, return true. if not, use this new gcd to compare with next value
and create another gcd. This makes this algorithm to become O(N)
algorithm