[REVIEW] perfect squares.
Given a positive integer n, find the least number of perfect square
numbers (for example, 1, 4, 9, 16, ...) which sum to n.

Hint) How can you use dp to solve this problem? That is, which of the
subproblems will be of help to solve num_squares(n)?

Hint) How can you shift the dp algorithm into bfs?