[REVIEW] Daily Temperatures.
Given temperature per dates, return the gap between next shortest warmer
dates. That is,
if you're given [73, 74, 75, 71, 69, 72, 76, 73]
return [1, 1, 4, 2, 1, 1, 0, 0]

POINT) What values should you save, and what should be the first to
check to get the first warmer date compared to current date temperature?

ANSWR) If you store the dates of the ascending temperature only that is
past today, you can check from the first until you get the date that is
warmer than temperature of today.
 Therefore, remove the temperature that is colder than today and put the
current date at the top of the stack. For example, if you have

3rd, 4th, 5th, 6th, 7th, 8th
[73, 68, 79, 71, 53, 88] and if you'll be checking the next warmer date of
2nd, only dates you'll be needing is

3rd, 5th, 8th. [73, 79, 88]