[REVIEW]
1. noi

no need to review. It's simply the replica but done in dfs.

2. reverse_polish_notation.cpp

return 2 + 3 when given {2, 3, + }

POINT) use of unordered_set to mark operators

3. clone graph.

when given with bidirectional graph, copy it. Node is given as

class Node {
public:
    int val;
    std::vector<Node*> neighbors;

    Node() {}

    Node(int _val, std::vector<Node*> _neighbors) {
	val = _val;
	neighbors = _neighbors;
    }
};

POINT) create unordered map of <node*, node*>. key is the node to copy.

POINT) at each step, neighbor nodes are added to copied node inside the
cloning graph function.