[REVIEW] Weekly contest 162
1252, 1253, 1254 are all simple simulations.

1255. Maximum Score Words Formed by Letters

POINT) How to use backtracking to go through all the words.

ANSWR) Run summing up the score of the case whether you're making the
ith word or not.
Before running previous process.
Check if ith word is able to be created. If it is, create it and
go through the process. After it's finished, recover the count of
letters used.
-> the reason why this problem is of backtracking